Longitude / Latidude Fix

+++ DroneApplication/src/main/java/br/fiap/integrations/DroneApplication/dtos/DroneDTO.java
@@ -10,9 +10,9 @@ public class DroneDTO {
     @NotNull
     private double temperature;
     @NotNull
-    private BigDecimal latitude;
+    private Double latitude;
     @NotNull
-    private BigDecimal longitude;
+    private Double longitude;
     @NotNull
     private double humidity;
     private boolean tracker;
@@ -33,19 +33,19 @@ public class DroneDTO {
         this.temperature = temperature;
     }
 
-    public BigDecimal getLatitude() {
+    public Double getLatitude() {
         return latitude;
     }
 
-    public void setLatitude(BigDecimal latitude) {
+    public void setLatitude(Double latitude) {
         this.latitude = latitude;
     }
 
-    public BigDecimal getLongitude() {
+    public Double getLongitude() {
         return longitude;
     }
 
-    public void setLongitude(BigDecimal longitude) {
+    public void setLongitude(Double longitude) {
         this.longitude = longitude;
     }
